pyproject.toml
[project]
name = "ict-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=1.0.4",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-pinecone>=0.1.0",
    "chromadb>=0.5.0",
    "pinecone-client>=3.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
]
